refactor: move chat config handlers into chats.go
